refactor(launcher): Rename bs-config -> bootstrap-config
Aligns naming and avoids (personal) confusion
